Delete controller interfaces
