+++ src/main/kotlin/es/uniovi/apuntesuniovi/controllers/commands/roles/SaveRole.kt
@@ -1,18 +1,21 @@
-package es.uniovi.apuntesuniovi.controllers.impl.roles
+package es.uniovi.apuntesuniovi.controllers.commands.roles
 
+import com.google.gson.Gson
 import es.uniovi.apuntesuniovi.infrastructure.Command
 import es.uniovi.apuntesuniovi.log.LogService
 import es.uniovi.apuntesuniovi.servicies.RoleService
 import es.uniovi.apuntesuniovi.servicies.dtos.entities.RoleDto
 
-class FindAllRoles(
-        private val roleService: RoleService
+class SaveRole(
+        private val roleService: RoleService,
+        private val json: String
 ) : Command<List<RoleDto>> {
     private val logService = LogService(this.javaClass)
 
     override fun execute(): List<RoleDto> {
         logService.info("execute() - start")
-        val result = roleService.findAll()
+        val roleDto = Gson().fromJson(json, RoleDto::class.java)
+        val result = roleService.save(roleDto)
         logService.info("execute() - end")
         return result
     }
