+++ src/main/kotlin/es/uniovi/apuntesuniovi/controllers/commands/subjects/FindAllSubjects.kt
@@ -1,4 +1,4 @@
-package es.uniovi.apuntesuniovi.controllers.impl.subjects
+package es.uniovi.apuntesuniovi.controllers.commands.subjects
 
 import es.uniovi.apuntesuniovi.infrastructure.Command
 import es.uniovi.apuntesuniovi.log.LogService
