+++ src/main/kotlin/es/uniovi/apuntesuniovi/infrastructure/constants/SecurityConstants.kt
@@ -1,12 +1,13 @@
-package es.uniovi.apuntesuniovi.servicies.security
+package es.uniovi.apuntesuniovi.infrastructure.constants
 
+/**
+ * Defines the constants of security services
+ */
 object SecurityConstants {
-    // Spring Security
     const val LOGIN_URL = "/login"
     const val HEADER_STRING = "Authorization"
     const val TOKEN_BEARER_PREFIX = "Bearer "
 
-    // JWT
     const val SECRET = "SecretKeyToGenJWTs"
     private const val HOURS = 6
     const val EXPIRATION_TIME = 1000 * 60 * 60 * HOURS
