+++ src/main/kotlin/es/uniovi/apuntesuniovi/servicies/SubjectService.kt
@@ -2,7 +2,19 @@ package es.uniovi.apuntesuniovi.servicies
 
 import es.uniovi.apuntesuniovi.servicies.dtos.entities.SubjectDto
 
+/**
+ * Service to manage subjects
+ */
 interface SubjectService {
+    /**
+     * Returns all subjects
+     */
     fun findAll(): List<SubjectDto>
+
+    /**
+     * Saves the subject
+     *
+     * @param subjectDto Subject to save
+     */
     fun save(subjectDto: SubjectDto): List<SubjectDto>
 }
