+++ src/main/kotlin/es/uniovi/apuntesuniovi/servicies/security/UserDetailsServiceImpl.kt
@@ -31,5 +31,4 @@ class UserDetailsServiceImpl @Inject constructor(
             throw UsernameNotFoundException(username)
         }
     }
-
 }
